Use consumable buffs more freely
If there is a (very) scary monster and we can't do anything except fight
it, use a buff potion if we have one.

I noticed the following behaviour: qw as DDFi ran into an orc warrior on
D:2 and tried to tab it to death without using the might potion in its
inventory.

This change causes qw to use potions much more frequently. Not as
frequently as berserk, but still quite often. Mostly qw finishes runs
with no spare consumables.